Added customer introduction
This adds a customer introduction that is used when the Plugin is first started, is not set to test mode or no secret key & merchant ID are used. The introduction acts as an overlay and has options to either add credentials, run the plugin in test mode or register for paytrail credentials at the paytrail website.